feat(glyphs): add reverse ! ? and missing v
Seems that somehow the Glyphs3 project file I've been using was older
and missing the 'v' glyph. It also didn't have the '%' glyph that Adrien
Chuttarsing made. I've updated the '%' glyph to be a combination of our
designs. I've added back in the missing 'v' glyph from the older font
files.

Also added the reverse ? and ! glyphs